Remove formatter module
